refactor: chatbot models
